finch-usage.ipynb: make code backward compat with old Finch server
This is so we can release new notebook before releasing new server.

Note the output is not backward compatible, only the code is.

So Jenkins will still be broken, but at least the notebook so not appear
broken to user when they run it from the tutorial-notebooks folder on
the Jupyter env.

See discussion in comment
https://github.com/bird-house/finch/pull/112#issuecomment-600224910