fix flakes8 failure E741 ambiguous variable name 'l'
$ flake8
./finch/processes/subset.py:40:36: E741 ambiguous variable name 'l'
./finch/processes/subset.py:46:35: E741 ambiguous variable name 'l'
./tests/locustfile.py:4:19: E741 ambiguous variable name 'l'
The command "flake8" exited with 1.

https://www.flake8rules.com/rules/E741.html

Do not use variables named 'I', 'O', or 'l' (E741)

Variables named I, O, and l can be very hard to read. This is because the letter I and the letter l are easily confused, and the letter O and the number 0 can be easily confused.

Change the names of these variables to something more descriptive.